Add option to hide summary of work item
The jsonresume schema indicates that work items have a summary prop.
I cannot tell if it is required.
In case it is required I am adding the option to hide the summary,
even if it exists in the defined JSON file.
This can be done by adding the hideSummary property to a work item,
and giving it a truthy value.
This addition is allowed by the schema.